initial implementation of fd based loop

+++ src/heron_rdma_client.cpp
@@ -9,7 +9,7 @@ RDMAClient::RDMAClient(RDMAOptions *opts, RDMAFabric *rdmaFabric, RDMAEventLoop 
   Init();
 }
 
-RDMAClient::RDMAClient(RDMAOptions *opts, RDMAFabric *rdmaFabric, RDMADatagram *loop, uint32_t target_id)
+RDMAClient::RDMAClient(RDMAOptions *opts, RDMAFabric *rdmaFabric, RDMADatagram *loop, uint16_t target_id)
     : RDMABaseClient(opts, rdmaFabric, loop, target_id) {
   Init();
 }
